Added basic weapons, still have to implement them and spawn projectiles

+++ src/comp_shipparams.cpp
@@ -7,7 +7,7 @@
 #include "../../include/sprite.h"
 
 CCompShipParams::CCompShipParams(CEntity * et, uint16 linear, uint16 angular,
-uint16 energy, uint16 hitpoints):
+float energy, uint16 hitpoints):
 	CComponent(et), m_linearSpeed(linear), m_angularSpeed(angular),
 	m_energy(energy), m_hitPoints(hitpoints) {}
